src/requirements.txt
numpy==2.1.2
pygame-ce==2.5.2
pyscroll==2.31
python-i18n==0.3.9
PyTMX==3.32
